Added current user verification

+++ FMI-Materials/src/main/java/com/fmi/materials/service/impl/CourseListServiceImpl.java
@@ -8,11 +8,13 @@ import com.fmi.materials.dto.courselist.CourseListDto;
 import com.fmi.materials.dto.courselist.CourseListDtoWithId;
 import com.fmi.materials.exception.EntityAlreadyExistsException;
 import com.fmi.materials.exception.EntityNotFoundException;
+import com.fmi.materials.exception.InvalidArgumentException;
 import com.fmi.materials.mapper.CourseDtoMapper;
 import com.fmi.materials.mapper.CourseListDtoMapper;
 import com.fmi.materials.mapper.UserDtoMapper;
 import com.fmi.materials.model.Course;
 import com.fmi.materials.model.CourseList;
+import com.fmi.materials.model.CustomUserDetails;
 import com.fmi.materials.model.User;
 import com.fmi.materials.repository.CourseListRepository;
 import com.fmi.materials.repository.CourseRepository;
@@ -21,6 +23,8 @@ import com.fmi.materials.service.UserService;
 import com.fmi.materials.vo.ExceptionMessage;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -50,12 +54,27 @@ public class CourseListServiceImpl implements CourseListService {
         this.courseDtoMapper = courseDtoMapper;
     }
 
+    private Boolean authenticateCurrentUser(Long userId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        CustomUserDetails userDetails = (CustomUserDetails)authentication.getPrincipal();
+        Long loggedUserId = userDetails.getId();
+
+        if(loggedUserId != userId) {
+            return false;
+        }
+        return true;
+    }
+
     @Override
     public CourseListDtoWithId createCourseList(CourseListDto courseListDto, Long userId) {
         if (this.courseListRepository.findByListName(courseListDto.getListName()) != null) {
             throw new EntityAlreadyExistsException(ExceptionMessage.ALREADY_EXISTS.getFormattedMessage("CourseList", "name", courseListDto.getListName()));
         }
         try {
+            if (!authenticateCurrentUser(userId)) {
+                throw new InvalidArgumentException(ExceptionMessage.INVALID_OPERATION.getFormattedMessage());
+            }
+
             User user = this.userDtoMapper.convertToEntityWithId(this.userService.findUserById(userId));
             CourseList courseList = this.courseListDtoMapper.convertToEntity(courseListDto);
             courseList.setUser(user);
@@ -67,7 +86,11 @@ public class CourseListServiceImpl implements CourseListService {
     }
 
     @Override
-    public CourseListDtoWithId updateCourseList(CourseListDtoWithId courseListDtoWithId) {
+    public CourseListDtoWithId updateCourseList(Long userId, CourseListDtoWithId courseListDtoWithId) {
+        if (!authenticateCurrentUser(userId)) {
+            throw new InvalidArgumentException(ExceptionMessage.INVALID_OPERATION.getFormattedMessage());
+        }
+
         if (!this.courseListRepository.existsById(courseListDtoWithId.getId())) {
             throw new EntityNotFoundException(ExceptionMessage.NOT_FOUND.getFormattedMessage("CourseList", "id", courseListDtoWithId.getId()));
         }
@@ -78,6 +101,10 @@ public class CourseListServiceImpl implements CourseListService {
 
     @Override
     public void deleteCourseList(Long userId, Long courseListId) {
+        if (!authenticateCurrentUser(userId)) {
+            throw new InvalidArgumentException(ExceptionMessage.INVALID_OPERATION.getFormattedMessage());
+        }
+
         if (this.getCourseList(courseListId, userId) != null) {
             this.courseListRepository.deleteById(courseListId);
         }
@@ -107,6 +134,10 @@ public class CourseListServiceImpl implements CourseListService {
     @Override
     public CourseListDtoWithId addCourseToList(Long courseId, Long courseListId, Long userId) {
         try {
+            if (!authenticateCurrentUser(userId)) {
+                throw new InvalidArgumentException(ExceptionMessage.INVALID_OPERATION.getFormattedMessage());
+            }
+
             Course course = this.courseRepository.findById(courseId)
                 .orElseThrow(() -> new EntityNotFoundException(ExceptionMessage.NOT_FOUND.getFormattedMessage("CourseList", "id", courseListId)));
 
@@ -128,6 +159,10 @@ public class CourseListServiceImpl implements CourseListService {
     @Override
     public void deleteCourseFromCourseList(Long userId, Long courseListId, Long courseId) {
         try {
+            if (!authenticateCurrentUser(userId)) {
+                throw new InvalidArgumentException(ExceptionMessage.INVALID_OPERATION.getFormattedMessage());
+            }
+
             CourseList courseList = this.courseListDtoMapper.convertToEntityWithId(this.getCourseList(courseListId, userId));
             User user = this.userDtoMapper.convertToEntityWithId(this.userService.findUserById(userId));
             courseList.setUser(user);
